requirements.txt
faiss-cpu==1.8.0
sentence-transformers==4.1.0
torch==2.2.2
numpy
scipy
matplotlib
pytest
